VZ-7362: Move workload ready check to a common area
