+++ pkg/k8s/ready/daemonset_ready.go
@@ -1,12 +1,11 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package status
+package ready
 
 import (
 	"context"
 	"fmt"
-	"github.com/verrazzano/verrazzano/pkg/k8s/status"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	appsv1 "k8s.io/api/apps/v1"
@@ -66,7 +65,7 @@ func DaemonSetsAreReady(log vzlog.VerrazzanoLogger, client client.Client, namesp
 // running and ready
 func podsReadyDaemonSet(log vzlog.VerrazzanoLogger, client client.Client, namespacedName types.NamespacedName, selector *metav1.LabelSelector, expectedNodes int32, prefix string) bool {
 	// Get a list of pods for a given namespace and labels selector
-	pods := status.GetPodsList(log, client, namespacedName, selector)
+	pods := GetPodsList(log, client, namespacedName, selector)
 	if pods == nil {
 		return false
 	}
@@ -111,7 +110,7 @@ func podsReadyDaemonSet(log vzlog.VerrazzanoLogger, client client.Client, namesp
 	}
 
 	// Make sure pods using the latest controllerRevision resource are ready.
-	podsReady, success := status.EnsurePodsAreReady(log, savedPods, expectedNodes, prefix)
+	podsReady, success := EnsurePodsAreReady(log, savedPods, expectedNodes, prefix)
 	if !success {
 		return false
 	}
