+++ pkg/k8s/ready/ingresses_present.go
@@ -1,6 +1,6 @@
 // Copyright (c) 2021, 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-package status
+package ready
 
 import (
 	"context"
