+++ platform-operator/controllers/verrazzano/component/console/console.go
@@ -6,7 +6,7 @@ package console
 import (
 	"fmt"
 	"github.com/verrazzano/verrazzano/pkg/bom"
-	"github.com/verrazzano/verrazzano/pkg/k8s/status"
+	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
@@ -67,7 +67,7 @@ func AppendOverrides(ctx spi.ComponentContext, _ string, _ string, _ string, kvs
 }
 
 func isConsoleReady(ctx spi.ComponentContext) bool {
-	return status.DeploymentsAreReady(
+	return ready.DeploymentsAreReady(
 		ctx.Log(),
 		ctx.Client(),
 		[]types.NamespacedName{
