+++ platform-operator/controllers/verrazzano/component/grafana/grafana.go
@@ -6,7 +6,7 @@ package grafana
 import (
 	"fmt"
 
-	"github.com/verrazzano/verrazzano/pkg/k8s/status"
+	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"k8s.io/apimachinery/pkg/types"
@@ -18,7 +18,7 @@ const grafanaDeployment = "vmi-system-grafana"
 func isGrafanaInstalled(ctx spi.ComponentContext) bool {
 	prefix := newPrefix(ctx.GetComponent())
 	deployments := newDeployments()
-	return status.DoDeploymentsExist(ctx.Log(), ctx.Client(), deployments, 1, prefix)
+	return ready.DoDeploymentsExist(ctx.Log(), ctx.Client(), deployments, 1, prefix)
 }
 
 // isGrafanaReady checks that the deployment has the minimum number of replicas available and
@@ -26,7 +26,7 @@ func isGrafanaInstalled(ctx spi.ComponentContext) bool {
 func isGrafanaReady(ctx spi.ComponentContext) bool {
 	prefix := newPrefix(ctx.GetComponent())
 	deployments := newDeployments()
-	return status.DeploymentsAreReady(ctx.Log(), ctx.Client(), deployments, 1, prefix) && common.IsGrafanaAdminSecretReady(ctx)
+	return ready.DeploymentsAreReady(ctx.Log(), ctx.Client(), deployments, 1, prefix) && common.IsGrafanaAdminSecretReady(ctx)
 }
 
 // newPrefix creates a component prefix string
