+++ platform-operator/controllers/verrazzano/component/istio/istio_component.go
@@ -14,8 +14,8 @@ import (
 	ctrlerrors "github.com/verrazzano/verrazzano/pkg/controller/errors"
 	"github.com/verrazzano/verrazzano/pkg/helm"
 	"github.com/verrazzano/verrazzano/pkg/istio"
+	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
 	"github.com/verrazzano/verrazzano/pkg/k8s/resource"
-	"github.com/verrazzano/verrazzano/pkg/k8s/status"
 	"github.com/verrazzano/verrazzano/pkg/k8s/webhook"
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
@@ -397,7 +397,7 @@ func (i istioComponent) IsReady(context spi.ComponentContext) bool {
 			Namespace: IstioNamespace,
 		},
 	}
-	ready := status.DeploymentsAreReady(context.Log(), context.Client(), deployments, 1, prefix)
+	ready := ready.DeploymentsAreReady(context.Log(), context.Client(), deployments, 1, prefix)
 	if !ready {
 		return false
 	}
