+++ platform-operator/controllers/verrazzano/component/mysqloperator/mysqloperator.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/verrazzano/verrazzano/pkg/bom"
-	"github.com/verrazzano/verrazzano/pkg/k8s/status"
+	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	installv1beta1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
@@ -56,12 +56,12 @@ func AppendOverrides(compContext spi.ComponentContext, _ string, _ string, _ str
 
 // isReady - component specific checks for being ready
 func isReady(ctx spi.ComponentContext) bool {
-	return status.DeploymentsAreReady(ctx.Log(), ctx.Client(), getDeploymentList(), 1, getPrefix(ctx))
+	return ready.DeploymentsAreReady(ctx.Log(), ctx.Client(), getDeploymentList(), 1, getPrefix(ctx))
 }
 
 // isInstalled checks that the deployment exists
 func isInstalled(ctx spi.ComponentContext) bool {
-	return status.DoDeploymentsExist(ctx.Log(), ctx.Client(), getDeploymentList(), 1, getPrefix(ctx))
+	return ready.DoDeploymentsExist(ctx.Log(), ctx.Client(), getDeploymentList(), 1, getPrefix(ctx))
 }
 
 func getPrefix(ctx spi.ComponentContext) string {
