+++ platform-operator/controllers/verrazzano/component/nginx/nginx.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/verrazzano/verrazzano/pkg/bom"
-	"github.com/verrazzano/verrazzano/pkg/k8s/status"
+	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	installv1beta1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 	vpoconst "github.com/verrazzano/verrazzano/platform-operator/constants"
@@ -50,7 +50,7 @@ func isNginxReady(context spi.ComponentContext) bool {
 	if err != nil && context.GetComponent() == ComponentName {
 		context.Log().Progressf("Ingress external IP pending for component %s: %s", ComponentName, err.Error())
 	}
-	return err == nil && status.DeploymentsAreReady(context.Log(), context.Client(), deployments, 1, prefix)
+	return err == nil && ready.DeploymentsAreReady(context.Log(), context.Client(), deployments, 1, prefix)
 }
 
 func AppendOverrides(context spi.ComponentContext, _ string, _ string, _ string, kvs []bom.KeyValue) ([]bom.KeyValue, error) {
