+++ platform-operator/controllers/verrazzano/component/prometheus/nodeexporter/nodeexporter.go
@@ -12,8 +12,8 @@ import (
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
 
+	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
-	"github.com/verrazzano/verrazzano/platform-operator/internal/k8s/status"
 	corev1 "k8s.io/api/core/v1"
 	netv1 "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -37,7 +37,7 @@ func isPrometheusNodeExporterReady(ctx spi.ComponentContext) bool {
 		},
 	}
 	prefix := fmt.Sprintf("Component %s", ctx.GetComponent())
-	return status.DaemonSetsAreReady(ctx.Log(), ctx.Client(), sets, 1, prefix)
+	return ready.DaemonSetsAreReady(ctx.Log(), ctx.Client(), sets, 1, prefix)
 }
 
 // PreInstall implementation for the Prometheus Node-Exporter Component
